Debug Menu and Move into home Added
In order to test move into home instead of having hardcoded values for
black home or white home (or fudging it so it works for now) I opt'd to
create a debug menu that allows you to change the state of the game by
clicking on the buttons in this menu. It should be noted The Dice
portion of this menu is not functional As I currently had no use for it
and didn't want to waste time on something we might not even use.

To bring this menu up simpily hit the 1 button on your keyboard(not num
pad). because this should not be used in release I chose to use
preprocessor statements to make it so if DEBUG_MENU is defined this code
will be added and if it isn't this code won't.

Now for the Move to home I made a few changes with the pieces mostly
just fixing up some bugs I had with the previous story of drag and drop
making it so you can only pick up pieces if they are not in home(off
board) and making sure that you dont have any pieces on the bar.

To be able to move your pieces into home a few conditions must be met
1. It must be your turn (right now ever piece is just player 1)
2. You must be in the move phase of your turn
3. all your piece must be in home set by in home variables

All that is left for this story is unit testing but I expect there to be
more manual testing then unit testing for this story because it's very
UI based. Ideally we would have 2 people review this story.

closes #23 Move into home
closes #55 Debug Menu